+++ docs/api.md
@@ -5,6 +5,7 @@
 ::: llmz.components.activations
 ::: llmz.components.transformers
 ::: llmz.datasets
+::: llmz.evaluate
 ::: llmz.generate
 ::: llmz.gpt2
 ::: llmz.train
